Shortcut text rendering when the font has no glyphs.
This is common with documents that have been through OCR software, and so
they have hidden text with the OCR results.  When rendering the document,
this text is all invisible, and so we can shortcut it.

Do this by adding isForHiddenText to the FontBoxFont and PDFontLike interfaces,
and implement it in all font subclasses (or their abstract base classes).